requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.5.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
et_xmlfile==2.0.0
extension==1.0.0
Flask==3.1.0
graphviz==0.20.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyGithub==2.6.1
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.2
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.2
